requirements.txt
Flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
requests==2.32.3
python-dateutil==2.9.0.post0
APScheduler==3.10.4
